Small tweaks to default components
iFixit's video player needs are much less than that of a company that
does full feature videos.  As of this version, there doesn't seem to be
an easy way to disable default components (ie SeekHandler) on
initialization.  Once this is easily/quickly changeable, i'll remove my
changes/comment-outs from the source, but until then this is the easiest
way.